f95693d0cef32fc124640a56737de09baaccaa0e Fix: SearchPane tables which have Ajax loaded language information were unclickable as the events weren't being added at the right point.
The `table().container()` element in Select was empty because we weren't
waiting for the table to be fully ready. Slightly hack fix is to just
set the selection style when the table is ready to make sure it has its
correct listeners.

https://datatables.net/forums/discussion/77161

8b5ce774e0f3e4b84c4e54eb5c076eb00302e3be Merge branch 'master' of github.com:DataTables/SearchPanes

48e15ea90919cf5f16154fd5bd04f6b378a9d30e Dev: Handle the sync case as well as async!

Sync to source repo @8b5ce774e0f3e4b84c4e54eb5c076eb00302e3be